// src/app/jobseeker/dashboard/components/jobseeker-profile/components/profile-project/model/projects.model.ts
export class Project {
    projectName:string;
    associatedWith: string;
    client: string;
    startDate: string;
    endDate: string;
    projectDescription: string;
    projectUrl: string; 
    location: string; 
    currentlyWorking: string; 
    empType: string; 
    workType: string; 
    projectSite: string; 
    role: string; 
    roleDescription: string; 
    skillsUsed: any = []; 
}